Helper

+++ Sistem-Manajemen-Bengkel/SMB_Backend/Model/RiwayatSparepartModel.cs
@@ -10,7 +10,7 @@ namespace Sistem_Manajemen_Bengkel.SMB_Backend.Model
     {
         public int id_riwayat_sparepart { get; set; }
         public int id_sparepart { get; set; }
-        public decimal jumlah { get; set; }
+        public int jumlah { get; set; }
         public decimal harga { get; set; }
     }
 }
